+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/fsm/TransitionClassifier.scala
@@ -20,7 +20,7 @@ abstract class TransitionClassifier {
     * @param featureVector the feature vector to use to compute the distribution
     * @return a probability distribution over Transitions
     */
-  def getDistribution(featureVector: FeatureVector): Map[StateTransition, Double]
+  def getDistribution(featureVector: FeatureVector): Map[StateTransition, Float]
 
 }
